# packages/api-contract/openapi/boardops-v1.yaml
openapi: 3.1.0
info:
  title: BoardOps API
  version: 0.5.0
  description: Institution-scoped BoardOps API contract.
servers:
  - url: /api/v1
paths:
...
  /permissions/roles:
    get:
      operationId: listRoles
      description: Requires permissions.read.
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        "200":
          description: Institution roles and their permission assignments.
          content:
            application/json:
              schema:
                type: object
                required: [roles, requestId]
                properties:
                  roles:
                    type: array
                    items:
                      $ref: "#/components/schemas/PermissionRole"
                  requestId: { type: string }
        "403":
          $ref: "#/components/responses/ApiError"
    post:
      operationId: createRole
      description: Requires permissions.manage and recent STEP_UP verification.
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleWriteRequest"
      responses:
        "201":
          description: Custom role created.
          content:
            application/json:
              schema:
                type: object
                required: [role, requestId]
                properties:
                  role:
                    $ref: "#/components/schemas/PermissionRole"
                  requestId: { type: string }
        "4XX":
          $ref: "#/components/responses/ApiError"
  /permissions/roles/{roleId}:
    put:
      operationId: updateRole
      description: Requires permissions.manage and recent STEP_UP verification. System roles are immutable.
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RoleId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleWriteRequest"
      responses:
        "200":
          description: Custom role updated.
          content:
            application/json:
              schema:
                type: object
                required: [role, requestId]
                properties:
                  role:
                    $ref: "#/components/schemas/PermissionRole"
                  requestId: { type: string }
        "4XX":
          $ref: "#/components/responses/ApiError"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: boardops_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-session-token
  parameters:
...
    RoleId:
      name: roleId
      in: path
      required: true
      schema: { type: string }
  responses:
    ApiError:
      description: Structured API error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    PermissionRole:
      type: object
      required: [id, name, system, permissionCodes]
      properties:
        id: { type: string }
        name: { type: string }
        system: { type: boolean }
        permissionCodes:
          type: array
          uniqueItems: true
          items: { type: string }
        createdAtMs: { type: integer }
        updatedAtMs: { type: integer }
    RoleWriteRequest:
      type: object
      required: [name, permissionCodes]
      properties:
        name: { type: string, minLength: 2, maxLength: 80 }
        permissionCodes:
          type: array
          maxItems: 100
          uniqueItems: true
          items: { type: string, maxLength: 100 }
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            details: {}